Add translate-2.py
(Almost) drop-in replacement for translate.py that should work better
for large files until the issues with translate.py are fixed.

This code has not been thoroughly tested and may produce different
results from translate.py. A known issue is that it doesn't support
the --normalization_alpha parameter (it checks for -n in config file
instead and uses an alpha value of 1.0 if found or 0.0 otherwise).